requirements.txt
# Fact-Checking PoC Dependencies
streamlit>=1.32.0
google-genai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
